refactor for better orientation
