+++ backend/src/main/java/com/example/backend/security/AppUserDetailsService.java
@@ -1,5 +1,6 @@
 package com.example.backend.security;
 
+import com.example.backend.repo.AppUserRepo;
 import com.example.backend.security.model.AppUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.User;
