+++ backend/src/test/java/com/example/backend/controller/WorkoutControllerTest.java
@@ -2,7 +2,7 @@ package com.example.backend.controller;
 
 import com.example.backend.model.Workout;
 import com.example.backend.repo.WorkoutRepo;
-import com.example.backend.security.AppUserRepo;
+import com.example.backend.repo.AppUserRepo;
 import com.example.backend.security.model.AppUser;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
