+++ backend/src/test/java/com/example/backend/service/ExerciseServiceTest.java
@@ -5,7 +5,6 @@ import com.example.backend.model.ExerciseDTO;
 import com.example.backend.model.Progress;
 import com.example.backend.model.Sets;
 import com.example.backend.repo.ExerciseRepo;
-import com.example.backend.security.AppUserService;
 import com.example.backend.security.model.AppUserResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.security.core.userdetails.User;
